Extract folder browser entry models
